pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantag"
version = "0.5.0"
description = "Quantag Quantum Virtual Machine backend for Qiskit"
readme = "README.md"
authors = [
  { name = "Quantag IT Solutions GmbH", email = "info@quantag-it.com" }
]
license = { file = "LICENSE" }
keywords = ["quantum", "qiskit", "simulator", "cudaq"]
requires-python = ">=3.8"
dependencies = [
  "qiskit>=1.0.0,<2.0.0",
  "qiskit-optimization>=0.6.0",
  "pandas>=2.0.0",
  "requests>=2.0.0",
  "numpy>=1.20",
  "Pillow>=9.0"
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering :: Physics"
]

[project.optional-dependencies]
dwave = [
  "dwave-ocean-sdk>=6.5.0"
]

[project.urls]
Homepage = "https://quantum.quantag-it.com"
Source = "https://github.com/quantag/quantag-python"

